+++ src/controller.js
@@ -1,8 +1,12 @@
 /* eslint-disable no-plusplus */
 const fs = require('fs');
 const path = require('path');
-const util = require('util');
-const { task1: filterGoods, task2: goodsWithMaxCost, task3 } = require('./task');
+const {
+  task1: filterGoods,
+  task2: goodsWithMaxCost,
+  task3,
+  task4: { myMap, myMapAsync, getDiscountCallback, getDiscountPromise, getDiscountAsyncAwait },
+} = require('./task');
 
 const pathToFile = path.resolve(__dirname, '../', 'goods.json');
 
@@ -69,43 +73,6 @@ function newData(data, response) {
 
 // Homework-03
 
-function myMap(array, callback) {
-  const result = [];
-  for (let index = 0; index < array.length; index++) {
-    result.push(callback(array[index]));
-  }
-  return result;
-}
-
-async function myMapAsync(array, callback) {
-  const result = [];
-  // eslint-disable-next-line no-restricted-syntax
-  for await (const element of array) {
-    result.push(await callback(element));
-  }
-  return result;
-}
-
-function generateDiscount(callback) {
-  setTimeout(() => {
-    const discount = Math.floor(Math.random() * 99 + 1);
-    if (discount < 20) callback(null, discount);
-    else callback(new Error('The store cannot afford such big discounts.'));
-  }, 50);
-}
-
-const generateDiscountPromisified = util.promisify(generateDiscount);
-
-function generateDiscountPromise() {
-  return new Promise((resolve, reject) =>
-    generateDiscount((error, discount) => (error ? reject(error) : resolve(discount))),
-  );
-}
-
-function getDiscountCallback(callback) {
-  generateDiscount((err, discount) => (err ? getDiscountCallback(callback) : callback(discount)));
-}
-
 function discountCallback(response) {
   const standard = task3(readFileStorage(response));
   let mapped = 0;
@@ -149,23 +116,6 @@ function discountCallback(response) {
   }
 }
 
-function recursionPromise(times, discounts, callback) {
-  generateDiscountPromisified()
-    .then((discount) => {
-      discounts.push(discount);
-      if (discounts.length < times) recursionPromise(times, discounts, callback);
-      else callback(discounts);
-    })
-    .catch(() => recursionPromise(times, discounts, callback));
-}
-
-function getDiscountPromise(times, callback) {
-  const discounts = [];
-  for (let counter = 0; counter < times; counter++) {
-    recursionPromise(times, discounts, callback);
-  }
-}
-
 function discountPromise(response) {
   const standard = task3(readFileStorage(response));
   let mapped = 0;
@@ -195,14 +145,6 @@ function discountPromise(response) {
   function sendResponse() {
     response.write(JSON.stringify(discountedGoods));
     response.end();
-  }
-}
-
-async function getDiscountAsyncAwait() {
-  try {
-    return await generateDiscountPromise();
-  } catch (err) {
-    return getDiscountAsyncAwait();
   }
 }
