+++ src/task/index.js
@@ -1,5 +1,6 @@
 const task1 = require('./task1');
 const task2 = require('./task2');
 const task3 = require('./task3');
+const task4 = require('./task4');
 
-module.exports = { task1, task2, task3 };
+module.exports = { task1, task2, task3, task4 };
